% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pad_to_sample_id.R
\name{get_pad_to_sample_id}
\alias{get_pad_to_sample_id}
\title{Get Sample IDs from PAD IDs in Metadata}
\usage{
get_pad_to_sample_id(
  these_pad_ids = NULL,
  this_metadata = NULL,
  return_all = FALSE
)
}
\arguments{
\item{these_pad_ids}{Character vector of PAD IDs to match.}

\item{this_metadata}{Data frame containing metadata with columns \code{sample_id}, \code{tma_pad_id}, and \code{sample_category_group}.}
}
\value{
Character vector of sample IDs matching the requested PAD IDs and quality criteria.
}
\description{
This function returns the sample IDs corresponding to provided PAD IDs from a metadata data frame.
It filters for unique urothelial carcinoma (UC) samples of high or low quality.
}
\examples{
#get metadata
test_meta = usqUtils::get_usq_metadata(return_this = "full_meta")

#return sample IDs for the provided PAD IDs
get_pad_to_sample_id(these_pad_ids = c("PH_4222_20", "PM_7813_20"),
                     this_metadata = test_meta$usq_metadata_tidy)

}
